requirements.txt
sqlalchemy
redis
python-telegram-bot
aiogram
python-dotenv
psycopg2-binary
